simplify finish dataset and anonymizer
