--- src/renderer/src/components/finish-anonymizer/FinishAnonymizer.styles.ts
@@ -1,5 +1,0 @@
-import { styled } from "@/styles";
-
-export const Anchor = styled("a", {
-  mr: "auto",
-});
