--- src/renderer/src/components/finish-dataset/FinishDataset.styles.ts
@@ -1,8 +1,0 @@
-import { styled } from "@/styles";
-
-/**
- * @deprecated
- */
-export const Anchor = styled("a", {
-  mr: "auto",
-});
